fix: cleans up some code and tightens design some (#335)
* fix: tightens design on overview and reuse button everywhere

* fix: missing key in emoji list

* fix: link back to tjenester

* fix: proper routing and route naming

* refactor: extract page header

* fix: adds menu to tjenesteomrade pages

* fix: bug with text color buttons

* chore: removes unused typing, fixing types

* fix: build issues with logo from studentsplash

* fix: removes les mer om as it is implicit

* fix: makes ingress text consistent with other parts

* fix: mobile version menu

* fix: header on mobile

* fix: title for tjenesteområde to centered
