+++ src/components/button/button.module.css
@@ -1,20 +1,24 @@
 .button,
 .button:visited {
   border: 0;
-  background: var(--color-secondary1);
+  background: var(--bg, var(--color-secondary1));
   text-align: center;
   font-size: 1rem;
-  padding: 1rem 3rem;
-  color: var(--color-secondary1--text);
+  color: var(--color, var(--color-secondary1--text));
   border-radius: 0.5rem;
   font-weight: 500;
-  transition: background-color 250ms ease-out;
+  transition: background-color 250ms ease-out, scale 100ms ease-out;
   display: inline-block;
-  margin: auto;
+
+  border-radius: 4rem;
+  padding: 0.5rem 1rem;
 }
 .button:hover {
-  background-color: var(--color-secondary1__tint1);
-  color: var(--color-secondary1__tint1--text);
+  background-color: var(--hover, var(--color-secondary1__tint1));
+  color: var(--color, var(--color-secondary1__tint1--text));
+}
+.button:active {
+  scale: 0.99;
 }
 
 .buttonLink {
