+++ src/index/index.module.css
@@ -92,8 +92,6 @@
 
 .join_button {
   margin-top: 2rem;
-  border-radius: 4rem;
-  padding: 0.5rem 1rem;
 }
 
 @media (max-width: 500px) {
@@ -350,24 +348,5 @@
 
   .infoBlock__title {
     margin-left: 6rem;
-  }
-}
-
-.tjenesteomrader {
-  display: flex;
-  flex-direction: column;
-  gap: 3rem;
-  margin: 0 1rem;
-}
-
-@media (max-width: 700px) {
-  .tjenesteomrader__ingress {
-    text-align: center;
-  }
-}
-
-@media (max-width: 400px) {
-  .tjenesteomrader .join_title {
-    font-size: 3rem;
   }
 }
