+++ src/tjenesteomrader/images/headerBackground/headerBackground.module.css
@@ -1,12 +1,11 @@
 .header__text {
   background-position-y: bottom;
-  background-size: cover;
-  padding-top: 5rem;
   padding: 5rem;
+  background-size: cover;
   height: 20rem;
 }
 
-.header__text > * {
+.header__heading {
   margin: 0;
 }
 
@@ -16,11 +15,11 @@
 
 @media (max-width: 500px) {
   .header__text {
-    padding-top: 11rem;
+    padding-top: 8rem;
     padding-inline: 3rem;
   }
 
-  .header__text h2 {
+  .header__text h1 {
     font-size: 1.875rem;
   }
 }
