+++ src/tjenesteomrader/tjenesteomrader.module.css
@@ -1,0 +1,31 @@
+.tjenesteomrader {
+  display: flex;
+  flex-direction: column;
+  gap: 3rem;
+  margin: 0 1rem;
+}
+.tjenesteomrader__content {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  max-width: 44rem;
+}
+.tjenesteomrader__ingress {
+  font-size: 1.2rem;
+}
+
+.title {
+  text-align: center;
+}
+
+@media (max-width: 700px) {
+  .tjenesteomrader__ingress {
+    text-align: center;
+  }
+}
+
+@media (max-width: 400px) {
+  .title {
+    font-size: 3rem;
+  }
+}
